+++ glean-core/python/glean/metrics/timing_distribution.py
@@ -114,9 +114,7 @@ class TimingDistributionMetricType:
         """
         return self._TimingDistributionContextManager(self)
 
-    def test_get_value(
-        self, ping_name: Optional[str] = None
-    ) -> Optional[DistributionData]:
+    def test_get_value(self, ping_name: Optional[str] = None) -> Optional[DistributionData]:
         """
         Returns the stored value for testing purposes only.