+++ glean-core/python/tests/metrics/test_event.py
@@ -23,9 +23,7 @@ ROOT = Path(__file__).parent
 
 
 class ClickKeys(metrics.EventExtras):
-    def __init__(
-        self, object_id: Optional[str] = None, other: Optional[str] = None
-    ) -> None:
+    def __init__(self, object_id: Optional[str] = None, other: Optional[str] = None) -> None:
         self._object_id = object_id
         self._other = other
 
@@ -322,20 +320,16 @@ def test_long_extra_values_record_an_error():
 
 
 def test_event_enum_is_generated_correctly():
-    metrics = load_metrics(
-        ROOT.parent / "data" / "core.yaml", config={"allow_reserved": True}
-    )
+    metrics = load_metrics(ROOT.parent / "data" / "core.yaml", config={"allow_reserved": True})
 
     metrics.environment.event_example.record(
         metrics.environment.EventExampleExtra(key1="value1", key2="value2")
     )
 
-    extra = metrics.environment.event_example.test_get_value()[0].extra
-    del extra['glean_timestamp']
     assert {
         "key1": "value1",
         "key2": "value2",
-    } == extra
+    } == metrics.environment.event_example.test_get_value()[0].extra
 
 
 def test_event_extra_is_generated_correctly():
@@ -348,12 +342,10 @@ def test_event_extra_is_generated_correctly():
         metrics.core.PreferenceToggledExtra(preference="value1", enabled=True)
     )
 
-    extra = metrics.core.preference_toggled.test_get_value()[0].extra
-    del extra['glean_timestamp']
     assert {
         "preference": "value1",
         "enabled": "true",
-    } == extra
+    } == metrics.core.preference_toggled.test_get_value()[0].extra
 
 
 def test_the_convenient_extrakeys_api():
@@ -406,9 +398,7 @@ def test_event_extra_does_typechecks():
     metrics.core.PreferenceToggledExtra(preference="value1")
     metrics.core.PreferenceToggledExtra(enabled=True)
     metrics.core.PreferenceToggledExtra(swapped=1)
-    extras = metrics.core.PreferenceToggledExtra(
-        preference="value1", enabled=True, swapped=1
-    )
+    extras = metrics.core.PreferenceToggledExtra(preference="value1", enabled=True, swapped=1)
     # Check conversion to FFI types, extras are sorted by name
     ffi = extras.to_ffi_extra()
     expected = {