+++ glean-core/python/tests/metrics/test_quantity.py
@@ -114,6 +114,4 @@ def test_negative_values_are_not_counted():
 
     # Check that the quantity was NOT recorded
     assert 1 == quantity_metric.test_get_value("store1")
-    assert 1 == quantity_metric.test_get_num_recorded_errors(
-        testing.ErrorType.INVALID_VALUE
-    )
+    assert 1 == quantity_metric.test_get_num_recorded_errors(testing.ErrorType.INVALID_VALUE)