+++ glean-core/python/tests/metrics/test_timing_distribution.py
@@ -105,9 +105,8 @@ def test_stopping_a_non_existent_timer_records_an_error():
         time_unit=TimeUnit.NANOSECOND,
     )
 
-    metric.stop_and_accumulate(TimerId(0))
-    assert 1 == metric.test_get_num_recorded_errors(
-        testing.ErrorType.INVALID_STATE)
+    metric.stop_and_accumulate(TimerId(id=0))
+    assert 1 == metric.test_get_num_recorded_errors(testing.ErrorType.INVALID_STATE)
 
 
 # Doesn't really test anything anymore